// src/app/models/DTO/BuyerCarDTO.ts
export class BuyerCarDTO {
    id: Number = null;
    carSegment: String = '';
    mark: String = '';
    model: String = '';
    body: String = '';
    trunkCapacity: Number = null;
    drive: String = '';
    engine: String = '';
    acceleration: String = '';
    horsePower: String = '';
    fuelConsumption: String = '';
    price: String = '';
    date: String = '';
}